Added RPS.cmm and Added Funcs
Added a rock paper scissors program made with c-- (.cmm). Also added the getRand function that takes a min and max and sets active to it. Also added &= which takes the number of ifs ahead of it to allow stacking of them (also would work for multiple lines within an if).